Retype RoutineContext (6502 A/X/Y + flags) to u8
First swing of the u8 reimplementation: the register/flag carrier becomes u8
(it always held byte values; overflow-checks=false makes u8 arithmetic wrap
like the NES). Boundary casts to/from the still-i32 fields/locals were inserted
mechanically from rustc's own type-mismatch spans (304 sites) plus a few manual
fixes; many dissolve as fields migrate to u8 next. Verified byte-exact:
800-frame RAM trace identical to baseline, 34 test bins green, 240f smoke clean.

+++ src/engine.rs
@@ -52,13 +52,13 @@ pub mod reg {
 
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
 pub struct RoutineContext {
-    pub value: i32,
-    pub index: i32,
-    pub offset: i32,
-    pub carry: i32,
-    pub zero: i32,
-    pub negative: i32,
-    pub overflow: i32,
+    pub value: u8,
+    pub index: u8,
+    pub offset: u8,
+    pub carry: u8,
+    pub zero: u8,
+    pub negative: u8,
+    pub overflow: u8,
 }
 
 pub type RoutineFn = fn(&mut Engine, &mut RoutineContext);
